refactor(channels): own inbound envelope compat in core
